% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hof.R
\name{hof}
\alias{hof}
\title{Create a higher order function}
\usage{
hof(fn)
}
\arguments{
\item{fn}{The original function.}
}
\value{
A function with class 'hof', wrapping the original.
}
\description{
Create a higher order function
}
\examples{

}
