% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hof.R
\name{print.hof}
\alias{print.hof}
\title{Print a hof object}
\usage{
\method{print}{hof}(x, ...)
}
\arguments{
\item{x}{An object of class hof.}

\item{...}{Unused.}
}
\value{
NULL invisibly.
}
\description{
Print a hof object
}
